LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/script-src/manifest-v2-script-src-missing-scheme.html

<!DOCTYPE html><!-- webkit-test-runner [ contentSecurityPolicyExtensionMode=v2 ] -->
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="script-src 'self' localhost:8443 'nonce-test';">
    <p><br>Test that CSP for extensions blocks sources without a scheme.</p>
</head>
<body>
    <div id="result" text="FAIL">
      PASS
    </div>
    <script nonce='test'>
        if (window.testRunner) {
            testRunner.waitUntilDone();
            testRunner.dumpAsText();
        }

        let script = document.createElement('script');
        script.src = "https://localhost:8443/security/contentSecurityPolicy/resources/script.js";

        script.onload = function() {
            alert("FAIL: Should not have loaded script.");
            testRunner.notifyDone();
        };

        script.onerror = function() {
            alert("PASS: Successfully blocked script.");
            testRunner.notifyDone();
        };

        document.body.append(script);
    </script>
</body>
</html>
